Added blur background effect. Codified rounded artwork and shadow

+++ Airly/Objective-C Classes/SLColorArt/SLColorArt.h
@@ -31,7 +31,7 @@
 
 + (void)processImage:(UIImage *)image
         scaledToSize:(CGSize)scaleSize
-           threshold:(NSInteger)threshold
+           threshold:(CGFloat)threshold
           onComplete:(void (^)(SLColorArt *colorArt))completeBlock;
 
 @end
